feat: 파일 조회 경로 수정 및 Post.status enum 추가
- 파일 조회 경로 수정: 2025-08 -> 2025/08
- Post.status 추가 및 enum 'post_status'로 관리

+++ src/main/java/com/pyomin/cool/service/CategoryServiceImpl.java
@@ -28,7 +28,7 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     @Transactional(readOnly = true)
-    public Long getCategoryIdBySlug(String slug) {
+    public Integer getCategoryIdBySlug(String slug) {
         return categoryRepository.findBySlug(slug)
                 .orElseThrow(() -> new IllegalArgumentException("카테고리를 찾을 수 없습니다: " + slug))
                 .getId();
